Move GatewayConfigGroupsGenerator to PaymentBundle
