+++ src/Sylius/Bundle/PaymentBundle/DependencyInjection/Configuration.php
@@ -47,6 +47,15 @@ final class Configuration implements ConfigurationInterface
         $rootNode
             ->addDefaultsIfNotSet()
             ->children()
+                ->arrayNode('gateway_config')
+                    ->addDefaultsIfNotSet()
+                    ->children()
+                        ->arrayNode('validation_groups')
+                            ->useAttributeAsKey('name')
+                            ->variablePrototype()->end()
+                        ->end()
+                    ->end()
+                ->end()
                 ->scalarNode('driver')->defaultValue(SyliusResourceBundle::DRIVER_DOCTRINE_ORM)->end()
                 ->arrayNode('gateways')
                     ->useAttributeAsKey('name')
