+++ src/Sylius/Bundle/PayumBundle/DependencyInjection/Configuration.php
@@ -33,15 +33,6 @@ final class Configuration implements ConfigurationInterface
         $rootNode
             ->addDefaultsIfNotSet()
             ->children()
-                ->arrayNode('gateway_config')
-                    ->addDefaultsIfNotSet()
-                    ->children()
-                        ->arrayNode('validation_groups')
-                            ->useAttributeAsKey('name')
-                            ->variablePrototype()->end()
-                        ->end()
-                    ->end()
-                ->end()
                 ->scalarNode('driver')->defaultValue(SyliusResourceBundle::DRIVER_DOCTRINE_ORM)->end()
                 ->arrayNode('template')
                     ->addDefaultsIfNotSet()
