+++ src/Sylius/Bundle/PayumBundle/Resources/config/services/form.xml
@@ -12,11 +12,6 @@
 -->
 
 <container xmlns="http://symfony.com/schema/dic/services" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://symfony.com/schema/dic/services http://symfony.com/schema/dic/services/services-1.0.xsd">
-    <parameters>
-        <parameter key="sylius.form.type.gateway_config.validation_groups" type="collection">
-            <parameter>sylius</parameter>
-        </parameter>
-    </parameters>
     <services>
         <defaults public="true" />
